+++ frontend/src/app/video-viewer/video-viewer.ts
@@ -71,24 +71,32 @@ export default class VideoViewer {
     }
   }
 
-  public static elementInViewport(el) {
-    let top    = el.offsetTop;
-    let left   = el.offsetLeft;
-    let width  = el.offsetWidth;
-    let height = el.offsetHeight;
-
-    while (el.offsetParent) {
-      el = el.offsetParent;
-      top += el.offsetTop;
-      left += el.offsetLeft;
+  public static elementInViewport(elem) {
+    if (!(elem instanceof Element)) throw Error('DomUtil: elem is not an element.');
+    const style = getComputedStyle(elem);
+    if (style.display === 'none') return false;
+    if (style.visibility !== 'visible') return false;
+    // @ts-ignore
+    if (style.opacity < 0.1) return false;
+    // @ts-ignore
+    if (elem.offsetWidth + elem.offsetHeight + elem.getBoundingClientRect().height +
+      elem.getBoundingClientRect().width === 0) {
+      return false;
     }
-
-    return (
-      top < (window.pageYOffset + window.innerHeight) &&
-      left < (window.pageXOffset + window.innerWidth) &&
-      (top + height) > window.pageYOffset &&
-      (left + width) > window.pageXOffset
-    );
+    const elemCenter = {
+      x: elem.getBoundingClientRect().left + elem.offsetWidth / 2,
+      y: elem.getBoundingClientRect().top + elem.offsetHeight / 2
+    };
+    if (elemCenter.x < 0) return false;
+    if (elemCenter.x > (document.documentElement.clientWidth || window.innerWidth)) return false;
+    if (elemCenter.y < 0) return false;
+    if (elemCenter.y > (document.documentElement.clientHeight || window.innerHeight)) return false;
+    let pointContainer = document.elementFromPoint(elemCenter.x, elemCenter.y);
+    // @ts-ignore
+    do {
+      if (pointContainer === elem) return true;
+    } while (pointContainer && pointContainer === pointContainer.parentNode);
+    return false;
   }
 
   public static videoPlaying(el) {
